Tool_Calling/pyproject.toml
[project]
name = "tool-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.2.11",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
